Issue 67.2: extract neighbor-dispatch cluster from afxdp.rs
afxdp.rs goes from 1,275 → 1,086 prod LOC by relocating four
neighbor-related fns into a new sibling module
afxdp/neighbor_dispatch.rs:

- retry_pending_neigh (108 LOC) — post-poll loop that walks the
  per-binding pending-neighbor queue, re-issues bpf_fib_lookup +
  neighbor lookups, resumes flows whose neighbors resolved
- learn_dynamic_neighbor_from_packet (47 LOC) — RX-path entry point
  for learning a neighbor from an inbound ARP/NDP advert
- learn_dynamic_neighbor (26 LOC) — internal upsert helper
- build_missing_neighbor_session_metadata (16 LOC) — constructs the
  SessionMetadata stub used while neighbor is unresolved

Visibility:
- `retry_pending_neigh`, `learn_dynamic_neighbor_from_packet`,
  `build_missing_neighbor_session_metadata` are widened to
  `pub(super)` because they're called from afxdp.rs / poll_descriptor.rs.
- `learn_dynamic_neighbor` stays file-private (only called by
  `learn_dynamic_neighbor_from_packet` inside the new module).

afxdp.rs adds:
    mod neighbor_dispatch;
    use neighbor_dispatch::{
        build_missing_neighbor_session_metadata,
        learn_dynamic_neighbor_from_packet, retry_pending_neigh,
    };

Pure relocation — bodies byte-for-byte identical, all sibling-module
dependencies (parser, sharded_neighbor, types, etc.) reachable through
`use super::*;`.

cargo test --release: 865 passed, 0 failed.
